Rename package from ir to intermediate.
* ir can be a commonly used name for local variable and function
  parameters, which could potential shadow the package ir.
